Function for debugging optimization
gen_gauss_eval makes a function for debugging optimization based on
KL divergence. You give it two means and variances, and a selection
probability, and a sample size, and it creates a function which draws
samples from a gaussian mixture model for the given parameters, and
samples drawn from another GMM given in X, and compares their KL
divergence. Throw this created function as the evaluation function
for basinhopping and you should eventually find matching parameters
takes A LOT of iterations to find the parameters